Fix all warnings, yay!

+++ src/main/scala/LogicAnalyzer.scala
@@ -26,7 +26,7 @@ class LogicAnalyzer(dataWidth: Int, lineWidth: Int, samples: Int,
   val reqAddrWidth = log2Up(memDepth)
 
   // TODO: DRYify
-  val sIdle :: sArmed :: sRunning :: Nil = Enum(UInt(), 3)
+  val sIdle :: sArmed :: sRunning :: Nil = Enum(3)
   val stateWidth = log2Up(3)
 
   //
